Added RoomKey error and added test
