+++ components/ChoicesVoter.jsx
@@ -5,6 +5,7 @@ import Room from '../model/Room.js'
 import ChoicesView from './view/ChoicesView.jsx'
 import ErrorBoard from './ErrorBoard.jsx';
 import Loader from './Loader.jsx';
+import queryString from "query-string";
 
 
 let socket = false;
@@ -13,6 +14,8 @@ if (typeof socketAddress !== 'undefined') {
         socket = io(socketAddress);
     }
 }
+
+const params = queryString.parse(location.search);
 
 class Choices extends React.Component {
     constructor(props) {
@@ -23,7 +26,6 @@ class Choices extends React.Component {
         this.renderLoading = this.renderLoading.bind(this)
         this.BoardApi = this.props.BoardApi
         this.room = false;
-        this.timer = null;
         this.state = {
             leftCard: null,
             rightCard: null,
@@ -34,8 +36,7 @@ class Choices extends React.Component {
             selectedSide: null,
             roomVoters: [],
             boardId : null,
-            hasExtIdWrong:false,
-            timeout : false
+            hasUrlWrong:false,
         };
         if (this.props.roomKey !== null) {
             this.setUpRoom(component);
@@ -43,12 +44,6 @@ class Choices extends React.Component {
         if (socket) {
             this.setUpSocket(component);
         }
-    }
-
-    toggleTimeout () {
-        this.setState({
-            timeout: !this.state.timeout
-        })
     }
 
     setUpSocket(component) {
@@ -95,14 +90,21 @@ class Choices extends React.Component {
             })
         });
 
-        socket.on('extIdToVoters',function(extId){
+        socket.on('error', function () {
+            component.setState({
+                hasUrlWrong: true
+            })
+        });
+
+        socket.on('paramsToVoters',function(extId, fw){
+
             //if extId is null = object, if extId is string = string
             if(typeof(extId)!=="string"){
                 extId=JSON.stringify(extId);
             }
-           if(extId!==component.props.extId) {
+           if(extId!==component.props.extId || params.fw !== fw) {
                component.setState({
-                   hasExtIdWrong:true
+                   hasUrlWrong:true
                })
            }
         });
@@ -125,7 +127,7 @@ class Choices extends React.Component {
                     component.sortelloAvatar = '//www.gravatar.com/avatar/' + normalizedData.gravatar + '?s=64&d=identicon'
                 }
                 component.room.join(component.sortelloId, component.sortelloAvatar);
-                component.room.checkExtId();
+                component.room.checkParams();
                 component.room.castGetBoardId();
             }, function (e) {
                 console.log(e);
@@ -157,39 +159,26 @@ class Choices extends React.Component {
         return <ErrorBoard text="Url isn't correct, please contact board's administrator to get a new one."/>
     }
 
-    renderRoomKeyError(){
-        return <ErrorBoard text="RoomKey isn't correct, please contact board's administrator to get a new one."/>
-    }
-
-
     renderLoading() {
-        this.timer = setTimeout(this.toggleTimeout.bind(this), 5000);
         return <Loader/>;
     }
 
     render() {
         if ((this.state.leftCard == null || this.state.rightCard == null) && this.state.hasBoardPermissions === null &&
-        !this.state.hasExtIdWrong && !this.props.hasParamsMissing && !this.state.timeout) {
+        !this.state.hasUrlWrong && !this.props.hasParamsMissing) {
             return this.renderLoading()
         }
 
-        if (this.state.timeout){
-            clearTimeout(this.timer);
-            return this.renderRoomKeyError();
-        }
-
         if (this.state.hasBoardPermissions===false) {
-            clearTimeout(this.timer);
             return this.renderForbidden()
         }
 
-        if(this.state.hasExtIdWrong || this.props.hasParamsMissing){
-            clearTimeout(this.timer);
+        if(this.state.hasUrlWrong || this.props.hasParamsMissing){
+            //TODO hasParamsMissing unico variabile in props calcolata da App
             return this.renderUrlError();
         }
 
         if (this.state.ended) {
-            clearTimeout(this.timer);
             let url = this.props.BoardApi.getName() === "Trello"? "https://trello.com/b/" +this.state.boardId :
                 this.state.boardId+"#column-" + this.props.extId;
             return <PrioritizationEnd
